move "sign in as guest" button
