+++ Website/src/components/routes/SignIn.js
@@ -3,7 +3,7 @@ import { Link as RouterLink, useHistory } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import { setSnackbar } from 'src/actions/commonActions';
 import { grantAccess } from 'src/actions/authActions';
-import { signIn, signInOAuth2 } from 'src/actions/authApiActions';
+import { signIn, signUp, signInOAuth2 } from 'src/actions/authApiActions';
 import { saveState } from 'src/localStorage';
 import useForm from 'src/components/core/useForm';
 import Form from 'src/components/core/Form';
@@ -48,6 +48,15 @@ function SignIn() {
       }
     }
   );
+
+  const handleSubmitAnonymously = async () => {
+    try {
+      const data = await dispatch(signUp(inputs, true));
+      handleSignInSuccess(data, dispatch, history);
+    } catch (error) {
+      dispatch(setSnackbar({ text: 'Не удалось войти гостем' }));
+    }
+  };
 
   const handleSignInOAuth2 = useCallback(
     async response => {
@@ -98,6 +107,17 @@ function SignIn() {
         <EntryButtonContainer>
           <Button type="submit" color="primary" fullWidth large>
             Войти
+          </Button>
+        </EntryButtonContainer>
+        <EntryButtonContainer>
+          <Button
+            type="button"
+            color="default"
+            onClick={handleSubmitAnonymously}
+            fullWidth
+            large
+          >
+            Войти гостем
           </Button>
         </EntryButtonContainer>
         <EntryButtonContainer>
